requirements.txt
# Add your dependencies here
requests
python-dotenv
